+++ scripts/generate_enum_util.py
@@ -29,6 +29,7 @@ blacklist = [
     "ClaimState",
     "MoveBufferResult",
     "CSVBufferResidency",
+    "DuplicateEliminatedDomainCoverage",
     "AppendAdmission",
     "AppendReservationState",
     "BufferedPushState",
