+++ src/planner/subquery/delim_join_cte_rewriter.cpp
@@ -13,9 +13,9 @@
 #include "duckdb/main/query_profiler.hpp"
 #include "duckdb/main/settings.hpp"
 #include "duckdb/optimizer/column_binding_replacer.hpp"
-#include "duckdb/optimizer/duplicate_eliminated_domain_candidate.hpp"
-#include "duckdb/optimizer/duplicate_eliminated_domain_factorer.hpp"
-#include "duckdb/optimizer/duplicate_eliminated_domain_inliner.hpp"
+#include "duckdb/optimizer/duplicate_eliminated_domain/duplicate_eliminated_domain_candidate.hpp"
+#include "duckdb/optimizer/duplicate_eliminated_domain/duplicate_eliminated_domain_factorer.hpp"
+#include "duckdb/optimizer/duplicate_eliminated_domain/duplicate_eliminated_domain_inliner.hpp"
 #include "duckdb/planner/column_binding_map.hpp"
 #include "duckdb/planner/expression/bound_columnref_expression.hpp"
 #include "duckdb/planner/expression/bound_comparison_expression.hpp"
